Made enemy speed relative to scale

+++ Sprint 0/Scripts/Sprite/GoriyaBackSprite.cs
@@ -10,12 +10,12 @@ namespace Sprint_0.Scripts.Sprite
     {
         private Texture2D sprite;
         private Rectangle sourceRectangle;
-        private int scale;
+        private float scale;
 
         private SpriteEffects effect = SpriteEffects.None;
         private float framesPerSecond = 4;
         private float timeSinceFrame = 0;
-        public GoriyaBackSprite(Rectangle rectangle, int scale, Texture2D spriteSheet)
+        public GoriyaBackSprite(Rectangle rectangle, float scale, Texture2D spriteSheet)
         {
             sourceRectangle = rectangle;
             this.scale = scale;
@@ -40,7 +40,7 @@ namespace Sprint_0.Scripts.Sprite
         }
         public void Draw(SpriteBatch spriteBatch, Vector2 location)
         {
-            Rectangle destinationRectangle = new Rectangle((int)location.X, (int)location.Y, sourceRectangle.Width * scale, sourceRectangle.Height * scale);
+            Rectangle destinationRectangle = new Rectangle((int)location.X, (int)location.Y, (int)(sourceRectangle.Width * scale), (int)(sourceRectangle.Height * scale));
             spriteBatch.Draw(sprite, destinationRectangle, sourceRectangle, Color.White, 0, Vector2.Zero, effect, 0);
 
         }
